backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
redis>=5.0.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
chromadb>=0.4.22
google-generativeai>=0.3.2
gitpython>=3.1.41
langchain>=0.1.4
langchain-google-genai>=0.0.8
python-dotenv>=1.0.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
requests>=2.31.0
